Added audio

+++ Assets/Scripts/IHittable.cs
@@ -1,5 +1,5 @@
 
 public interface IHittable
 {
-    void OnHit(float damageDealt);
+    void OnHit(int damageDealt);
 }
